requirements.txt
bottle
simplejson
python-bitcoinrpc
git@github.com:OpenNetworking/gcoin-community.git